add docker scan test
