+++ xray_test.go
@@ -7,8 +7,6 @@ import (
 	coretests "github.com/jfrog/jfrog-cli-core/v2/utils/tests"
 	"github.com/jfrog/jfrog-cli-core/v2/xray/commands"
 	clientTestUtils "github.com/jfrog/jfrog-client-go/utils/tests"
-	"io/ioutil"
-	"os"
 	"os/exec"
 	"path/filepath"
 	"testing"
@@ -72,7 +70,7 @@ func initXrayCli() {
 func TestXrayBinaryScan(t *testing.T) {
 	initXrayTest(t, commands.GraphScanMinXrayVersion)
 	binariesPath := filepath.Join(filepath.FromSlash(tests.GetTestResourcesPath()), "xray", "binaries", "*")
-	output := runAuditCmdWithOutput(t, "scan", binariesPath, "--licenses", "--format=json")
+	output := xrayCli.RunCliCmdWithOutput(t, "scan", binariesPath, "--licenses", "--format=json")
 	verifyScanResults(t, output, 0, 1, 1)
 }
 
@@ -89,7 +87,7 @@ func TestXrayAuditNpm(t *testing.T) {
 	// Run npm install before executing jfrog xr npm-audit
 	assert.NoError(t, exec.Command("npm", "install").Run())
 
-	output := runAuditCmdWithOutput(t, "audit-npm", "--licenses", "--format=json")
+	output := xrayCli.RunCliCmdWithOutput(t, "audit-npm", "--licenses", "--format=json")
 	verifyScanResults(t, output, 0, 1, 1)
 }
 
@@ -103,7 +101,7 @@ func TestXrayAuditGradle(t *testing.T) {
 	prevWd := changeWD(t, tempDirPath)
 	defer clientTestUtils.ChangeDirAndAssert(t, prevWd)
 
-	output := runAuditCmdWithOutput(t, "audit-gradle", "--licenses", "--format=json")
+	output := xrayCli.RunCliCmdWithOutput(t, "audit-gradle", "--licenses", "--format=json")
 	verifyScanResults(t, output, 0, 0, 0)
 }
 
@@ -116,7 +114,7 @@ func TestXrayAuditMaven(t *testing.T) {
 	assert.NoError(t, fileutils.CopyDir(mvnProjectPath, tempDirPath, true, nil))
 	prevWd := changeWD(t, tempDirPath)
 	defer clientTestUtils.ChangeDirAndAssert(t, prevWd)
-	output := runAuditCmdWithOutput(t, "audit-mvn", "--licenses", "--format=json")
+	output := xrayCli.RunCliCmdWithOutput(t, "audit-mvn", "--licenses", "--format=json")
 	verifyScanResults(t, output, 0, 1, 1)
 }
 
@@ -142,28 +140,6 @@ func validateXrayVersion(t *testing.T, minVersion string) {
 func getXrayVersion() (version.Version, error) {
 	xrayVersion, err := xrayAuth.GetVersion()
 	return *version.NewVersion(xrayVersion), err
-}
-
-// Run `jfrog` command, redirect the stdout and return the output
-func runAuditCmdWithOutput(t *testing.T, args ...string) string {
-	newStdout, stdWriter, previousStdout := tests.RedirectStdOutToPipe()
-	// Restore previous stdout when the function returns
-	defer func() {
-		os.Stdout = previousStdout
-		assert.NoError(t, newStdout.Close())
-	}()
-	go func() {
-		err := xrayCli.Exec(args...)
-		assert.NoError(t, err)
-		// Closing the temp stdout in order to be able to read it's content.
-		assert.NoError(t, stdWriter.Close())
-	}()
-	content, err := ioutil.ReadAll(newStdout)
-	assert.NoError(t, err)
-	// Prints the redirected output to the standard output as well.
-	_, err = previousStdout.Write(content)
-	assert.NoError(t, err)
-	return string(content)
 }
 
 func verifyScanResults(t *testing.T, content string, minViolations, minVulnerabilities, minLicenses int) {
